# docs/rest/swagger.yaml
basePath: /api/v1
definitions:
...
  pr-reviewers-service_internal_generated_api_v1_handler.DummyLoginOut:
    properties:
      token:
        type: string
    type: object
  pr-reviewers-service_internal_generated_api_v1_handler.ErrorResponse:
    properties:
      error:
        properties:
          code:
            $ref: '#/definitions/pr-reviewers-service_internal_generated_api_v1_handler.ErrorResponseErrorCode'
          message:
            type: string
        type: object
    type: object
  pr-reviewers-service_internal_generated_api_v1_handler.ErrorResponseErrorCode:
    enum:
    - BAD_REQUEST
    - NO_CANDIDATE
    - NOT_ASSIGNED
    - NOT_FOUND
    - PR_EXISTS
    - PR_MERGED
    - TEAM_EXISTS
    - UNKNOWN
    type: string
    x-enum-varnames:
    - BADREQUEST
    - NOCANDIDATE
    - NOTASSIGNED
    - NOTFOUND
    - PREXISTS
    - PRMERGED
    - TEAMEXISTS
    - UNKNOWN
...
  pr-reviewers-service_internal_generated_api_v1_handler.MergePullRequestResponse:
    properties:
      pr:
        $ref: '#/definitions/pr-reviewers-service_internal_generated_api_v1_handler.PullRequest'
    type: object
...
  pr-reviewers-service_internal_generated_api_v1_handler.PostPullRequestMergeJSONRequestBody:
    properties:
      pull_request_id:
        type: string
    required:
    - pull_request_id
    type: object
...
  pr-reviewers-service_internal_generated_api_v1_handler.PostTeamAddJSONRequestBody:
    properties:
      members:
        items:
          $ref: '#/definitions/pr-reviewers-service_internal_generated_api_v1_handler.TeamMember'
        type: array
      team_name:
        type: string
    required:
    - members
    - team_name
    type: object
...
  pr-reviewers-service_internal_generated_api_v1_handler.PullRequest:
    properties:
      assigned_reviewers:
        description: AssignedReviewers user_id назначенных ревьюверов (0..2)
        items:
          type: string
        type: array
      author_id:
        type: string
      created_at:
        type: string
      merged_at:
        type: string
      pull_request_id:
        type: string
      pull_request_name:
        type: string
      status:
        $ref: '#/definitions/pr-reviewers-service_internal_generated_api_v1_handler.PullRequestStatus'
    required:
    - assigned_reviewers
    - author_id
    - pull_request_id
    - pull_request_name
    - status
    type: object
...
  pr-reviewers-service_internal_generated_api_v1_handler.PullRequestStatus:
    enum:
    - MERGED
    - OPEN
    type: string
    x-enum-varnames:
    - PullRequestStatusMERGED
    - PullRequestStatusOPEN
...
  pr-reviewers-service_internal_generated_api_v1_handler.Team:
    properties:
      members:
        items:
          $ref: '#/definitions/pr-reviewers-service_internal_generated_api_v1_handler.TeamMember'
        type: array
      team_name:
        type: string
    required:
    - members
    - team_name
    type: object
  pr-reviewers-service_internal_generated_api_v1_handler.TeamMember:
    properties:
      is_active:
        type: boolean
      user_id:
        type: string
      username:
        type: string
    required:
    - user_id
    - username
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a service for working with PR Reviewers.
  title: PR Reviewers service
  version: "1.0"
paths:
  /dummyLogin:
    post:
      consumes:
      - application/json
      description: Get JWT token for testing purposes
      operationId: DummyLogin
      produces:
      - application/json
      responses:
        "200":
          description: Successfully logged in
          schema:
            $ref: '#/definitions/pr-reviewers-service_internal_generated_api_v1_handler.DummyLoginOut'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/pr-reviewers-service_internal_generated_api_v1_handler.ErrorResponse'
      summary: Dummy login
      tags:
      - User
...
  /pullRequest/merge:
    post:
      consumes:
      - application/json
      description: Merge an existing pull request
      operationId: MergePullRequest
      parameters:
      - description: Pull request merge data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/pr-reviewers-service_internal_generated_api_v1_handler.PostPullRequestMergeJSONRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: PR successfully merged or already merged
          schema:
            $ref: '#/definitions/pr-reviewers-service_internal_generated_api_v1_handler.MergePullRequestResponse'
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/pr-reviewers-service_internal_generated_api_v1_handler.ErrorResponse'
        "404":
          description: Pull request not found
          schema:
            $ref: '#/definitions/pr-reviewers-service_internal_generated_api_v1_handler.ErrorResponse'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/pr-reviewers-service_internal_generated_api_v1_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/pr-reviewers-service_internal_generated_api_v1_handler.ErrorResponse'
      summary: Merge pull request
      tags:
      - PullRequests
...
  /team/add:
    post:
      consumes:
      - application/json
      description: Create a new team with members (creates/updates users)
      operationId: AddTeam
      parameters:
      - description: Team data with members
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/pr-reviewers-service_internal_generated_api_v1_handler.PostTeamAddJSONRequestBody'
      produces:
      - application/json
      responses:
        "201":
          description: Team successfully created
          schema:
            $ref: '#/definitions/pr-reviewers-service_internal_generated_api_v1_handler.Team'
        "304":
          description: No changes - team exists and no users were changed or added
        "400":
          description: Validation failed or duplicate users
          schema:
            $ref: '#/definitions/pr-reviewers-service_internal_generated_api_v1_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/pr-reviewers-service_internal_generated_api_v1_handler.ErrorResponse'
      summary: Create team with members
      tags:
      - Teams
...
  /team/get:
    get:
      consumes:
      - application/json
      description: Get team details with members by team name
      operationId: GetTeam
      parameters:
      - description: Team name
        in: query
        name: team_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Team found successfully
          schema:
            $ref: '#/definitions/pr-reviewers-service_internal_generated_api_v1_handler.Team'
        "400":
          description: Invalid team_name parameter
          schema:
            $ref: '#/definitions/pr-reviewers-service_internal_generated_api_v1_handler.ErrorResponse'
        "404":
          description: Team not found
          schema:
            $ref: '#/definitions/pr-reviewers-service_internal_generated_api_v1_handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/pr-reviewers-service_internal_generated_api_v1_handler.ErrorResponse'
      summary: Get team information
      tags:
      - Teams
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"